cli: extract status read model helpers ( W4 )
